requirements.txt
Flask
picamera[array]
RPi.GPIO
